+++ src/plugins.ts
@@ -1,3 +1,5 @@
+export {css as cssLR} from '@codemirror/lang-css';
+export {javascript as javascriptLR} from '@codemirror/lang-javascript';
 export {css} from '@codemirror/legacy-modes/mode/css';
 export {javascript, json} from '@codemirror/legacy-modes/mode/javascript';
 export {lua} from '@codemirror/legacy-modes/mode/lua';